requirements.txt
evdev==1.4.0
MouseInfo==0.1.3
mss==6.1.0
numpy==1.21.0
opencv-python==4.5.2.54
paho-mqtt==1.5.1
Pillow==8.3.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.3
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-xlib==0.31
python3-xlib==0.15
PyTweening==1.0.3
six==1.16.0
